# Optimize memory allocations on building patterns with `strings.Builder`