--- a/zglob_test.go
+++ b/zglob_test.go
@@ -212,3 +212,24 @@ func TestGlobError(t *testing.T) {
 		t.Errorf(`zglob failed: expected %v but got %v`, nil, got)
 	}
 }
+
+func BenchmarkGlob(b *testing.B) {
+	tmpdir, savedCwd := setup()
+	defer os.RemoveAll(tmpdir)
+	defer os.Chdir(savedCwd)
+
+	for i := 0; i < b.N; i++ {
+		for _, test := range testGlobs {
+			if test.err != "" {
+				continue
+			}
+			got, err := Glob(test.pattern)
+			if err != nil {
+				b.Fatal(err)
+			}
+			if len(got) != len(test.expected) {
+				b.Fatalf(`zglob failed: pattern %q: expected %v but got %v`, test.pattern, test.expected, got)
+			}
+		}
+	}
+}