This PR contains the following improvements

- Optimize building patterns with `strings.Builder` (the primary change)
  - `strings.Builder` was added at Go 1.10 (the version in `go.mod` is 1.12)
- Add `BenchmarkGlob` to measure performance
- Small fixes to make staticcheck happy
- Fix a badge in readme

The performance gain by this PR is as follows:

Before:

```
$ go test -bench . -benchmem
goos: darwin
goarch: amd64
pkg: github.com/mattn/go-zglob
cpu: Intel(R) Core(TM) i9-10910 CPU @ 3.60GHz
BenchmarkGlob-20             236           5027318 ns/op          249671 B/op       4139 allocs/op
PASS
ok      github.com/mattn/go-zglob       2.004s
```

After:

```
$ go test -bench . -benchmem
goos: darwin
goarch: amd64
pkg: github.com/mattn/go-zglob
cpu: Intel(R) Core(TM) i9-10910 CPU @ 3.60GHz
BenchmarkGlob-20             238           5011009 ns/op          249391 B/op       3706 allocs/op
PASS
ok      github.com/mattn/go-zglob       1.823s
```

Summary:

- Allocations: 11.68% smaller
- CPU: 0.325% faster (identical)